--- boru_sample.conf
@@ -1,4 +1,0 @@
-group=wheel
-wrong_pw_sleep=1000
-session_ttl=5
-nopass=1
