+++ source/main.c
@@ -1,63 +1,24 @@
-#include <pwd.h>
-#include <grp.h>
-#include <err.h>
-#include <shadow.h>
-#include <crypt.h>
-#include <unistd.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "passwd.h"
-#include "sessions.h"
-
-#define VERSION "1.1.3:1"
-
-
-void getconf(FILE* fp, const char* entry, char* result, size_t len_result) {
-	char* line = NULL;
-	size_t len = 0;
-
-	fseek(fp, 0, SEEK_SET);
-
-	while (getline(&line, &len, fp) != -1) {
-		if (strncmp(entry, line, strlen(entry)) == 0) {
-			strtok(line, "=");
-			char* token = strtok(NULL, "=");
-			if (token) {
-				strncpy(result, token, len_result);
-				result[strcspn(result, "\n")] = 0;
-				free(line);
-				return;
-			}
-		}
-	}
-
-	errx(1, "Could not get '%s' entry in config", entry);
-}
-
-void runprog(char** program_argv) {
-	if (setuid(0) < 0)
-		err(1, "Could not setuid");
-	if (setgid(0) < 0)
-		err(1, "Could not setgid");
-
-	putenv("HOME=/root");
-
-	execvp(program_argv[0], program_argv);
-
-	err(1, program_argv[0]);
-}
+#include "common.h"
 
 int main(int argc, char** argv) {
 	char groupname[64], wrong_pw_sleep[64], session_ttl[64], nopass[64], password[128];
 	unsigned int sleep_us, tries, ts_ttl;
+	const char* config_path = "/etc/boru.conf";
 
-	if (argc == 1) {
-		printf("Bonek Root User version: %s\n\n", VERSION);
-		printf("Usage: %s [command]\n", argv[0]);
-		return 0;
-	}
+    if (argc == 1) {
+		printf(
+            " _               \n"
+            "| |_ ___ ___ _ _ \n"
+            "| . | . |  _| | |\n"
+            "|___|___|_| |___|\n\n"
+        );
+        printf("Version is %s!\n", VERSION);
+        printf("Usage: %s [command]\n", argv[0]);
+		if (access(config_path, F_OK) == -1) {
+            genconf(config_path);
+        }
+        return 0;
+    }
 
 	if (geteuid() != 0)
 		errx(1, "The boru binary needs to be installed as SUID.");
@@ -65,20 +26,16 @@ int main(int argc, char** argv) {
 	int ruid = getuid();
 	if (ruid == 0)
 		runprog(&argv[1]);
+	
+	FILE* fp = fopen(config_path, "r");
+    getconf(fp, "group", groupname, sizeof(groupname));
+    getconf(fp, "wrong_pw_sleep", wrong_pw_sleep, sizeof(wrong_pw_sleep));
+    getconf(fp, "session_ttl", session_ttl, sizeof(session_ttl));
+    getconf(fp, "nopass", nopass, sizeof(nopass));
+    sleep_us = atoi(wrong_pw_sleep) * 1000;
+    ts_ttl = atoi(session_ttl) * 60;
 
-	FILE* fp = fopen("/etc/boru.conf", "r");
-
-	if (!fp)
-		err(1, "Could not open /etc/boru.conf");
-
-	getconf(fp, "group", groupname, sizeof(groupname));
-	getconf(fp, "wrong_pw_sleep", wrong_pw_sleep, sizeof(wrong_pw_sleep));
-	getconf(fp, "session_ttl", session_ttl, sizeof(session_ttl));
-        getconf(fp, "nopass", nopass, sizeof(nopass));
-	sleep_us = atoi(wrong_pw_sleep) * 1000;
-	ts_ttl = atoi(session_ttl) * 60;
-
-	fclose(fp);
+    fclose(fp);
 
 	if (getsession(getppid(), ts_ttl, ruid) == 0)
 		runprog(&argv[1]);
